Rename and reformat gulpfile
